Renamed typo module. Speeding up pure fluid detection.
