+++ src/peregrinepy/writers/__init__.py
@@ -5,4 +5,4 @@ from .writeArbitraryArray import writeArbitraryArray
 from .writeDualTimeQnm1 import writeDualTimeQnm1
 from .writeConnectivity import writeConnectivity
 from .writeConfigFile import writeConfigFile
-from . import writerMetaData
+from . import writeMetaData
