+++ src/peregrinepy/writers/writeGrid.py
@@ -1,7 +1,7 @@
 import h5py
 import numpy as np
 from ..misc import progressBar
-from .writerMetaData import gridMetaData
+from .writeMetaData import gridMetaData
 
 
 def writeGrid(mb, path="./", precision="double", withHalo=False, lump=True):
